refactor: extract shared callback factory for auth and logout
Both oauthCallbackRoute and logoutCallbackRoute were identical
(check session, get originalRedirectUri, redirect with params).
Extract a createCallbackRoute(label) factory to avoid duplication.
